Revert "feat: Added support to revert from explorer (close #606) (#608)"
This reverts commit 02c983b78afa3cf4bb56e40f706e68e06ca1cc81.
