+++ src/commands/revert.ts
@@ -1,5 +1,5 @@
 import { SourceControlResourceState, window } from "vscode";
-import { checkAndPromptDepth, confirmRevert } from "../input/revert";
+import { SvnDepth } from "../common/types";
 import { Command } from "./command";
 
 export class Revert extends Command {
@@ -10,16 +10,32 @@ export class Revert extends Command {
   public async execute(...resourceStates: SourceControlResourceState[]) {
     const selection = await this.getResourceStates(resourceStates);
 
-    if (selection.length === 0 || !(await confirmRevert())) {
+    if (selection.length === 0) {
       return;
     }
 
+    const yes = "Yes I'm sure";
+    const answer = await window.showWarningMessage(
+      "Are you sure? This will wipe all local changes.",
+      { modal: true },
+      yes
+    );
+
+    if (answer !== yes) {
+      return;
+    }
+
+    const picks: any[] = [];
+
+    for (const depth in SvnDepth) {
+      if (SvnDepth.hasOwnProperty(depth)) {
+        picks.push({ label: depth, description: SvnDepth[depth] });
+      }
+    }
+
+    const placeHolder = "Select revert depth";
+    const pick = await window.showQuickPick(picks, { placeHolder });
     const uris = selection.map(resource => resource.resourceUri);
-    const depth = await checkAndPromptDepth(uris);
-
-    if (!depth) {
-      return;
-    }
 
     await this.runByRepository(uris, async (repository, resources) => {
       if (!repository) {
@@ -29,7 +45,7 @@ export class Revert extends Command {
       const paths = resources.map(resource => resource.fsPath);
 
       try {
-        await repository.revert(paths, depth);
+        await repository.revert(paths, pick.label);
       } catch (error) {
         console.log(error);
         window.showErrorMessage("Unable to revert");
