--- src/commands/revertExplorer.ts
@@ -1,41 +1,0 @@
-import { Uri, window } from "vscode";
-import { checkAndPromptDepth, confirmRevert } from "../input/revert";
-import { Command } from "./command";
-
-export class RevertExplorer extends Command {
-  constructor() {
-    super("svn.revertExplorer");
-  }
-
-  public async execute(_mainUri?: Uri, allUris?: Uri[]) {
-    if (!allUris) {
-      return;
-    }
-
-    const uris = allUris;
-    if (uris.length === 0 || !(await confirmRevert())) {
-      return;
-    }
-
-    const depth = await checkAndPromptDepth(uris);
-
-    if (!depth) {
-      return;
-    }
-
-    await this.runByRepository(uris, async (repository, resources) => {
-      if (!repository) {
-        return;
-      }
-
-      const paths = resources.map(resource => resource.fsPath);
-
-      try {
-        await repository.revert(paths, depth);
-      } catch (error) {
-        console.log(error);
-        window.showErrorMessage("Unable to revert");
-      }
-    });
-  }
-}
